refactor(core): move `groupInlineRuns` into `inline-runs.ts`
